% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_pred_summary.R
\name{create_obs_pred_summary}
\alias{create_obs_pred_summary}
\title{create_obs_pred_summary}
\usage{
create_obs_pred_summary(jags_dt_summary, jags_dat_model, jags_dat_prep)
}
\arguments{
\item{jags_dt_summary}{a data.table summary of the jags model}

\item{jags_dat_model}{the data that went into the jags model}

\item{jags_dat_prep}{the preperatory data that went into the jags model - purely so I can get the SEG information.}
}
\value{
data.frame
}
\description{
This function gets summary information for the predicted Y in the JAGS model, as I have laid it out, and combines it with the data so you can do really cool plots
}
