% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fig_four.R
\name{plot_fig_four}
\alias{plot_fig_four}
\title{plot_fig_four}
\usage{
plot_fig_four(obs_pred_summary)
}
\arguments{
\item{obs_pred_summary}{the summary data obtained from `create_obs_pred_summary.R`}
}
\value{
ggplot2 plot figure
}
\description{
plot figure four, which shows the observed and predicted values along with their credible intervals, colouring these according to whether the observed value lies within the 95% credible interval
}
